+++ src/IdentificationNumber.sln
@@ -1,0 +1,49 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33103.184
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "psafth.IdentificationNumber", "IdentificationNumber\psafth.IdentificationNumber.csproj", "{46E35B14-2DBB-4603-885B-A84A9D71BFC9}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "psafth.IdentificationNumber.Tests", "IdentificationNumber.Tests\psafth.IdentificationNumber.Tests.csproj", "{A06DF01A-8C46-410F-A10D-A97544E06FE5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "psafth.IdentificationNumber.Examples", "IdentificationNumber.Examples\psafth.IdentificationNumber.Examples.csproj", "{CB3087F6-70E0-41D9-A87F-025382234A8C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "psafth.IdentificationNumber.Swedish", "psafth.IdentificationNumber.Swedish\psafth.IdentificationNumber.Swedish.csproj", "{EA94EC6D-4F92-49C4-9C4E-A24200BF9DE6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "psafth.IdentificationNumber.Swedish.Tests", "psafth.IdentificationNumber.Swedish.Tests\psafth.IdentificationNumber.Swedish.Tests.csproj", "{77F256F5-BB49-4D7C-8D6C-5976D162727A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EA94EC6D-4F92-49C4-9C4E-A24200BF9DE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EA94EC6D-4F92-49C4-9C4E-A24200BF9DE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EA94EC6D-4F92-49C4-9C4E-A24200BF9DE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EA94EC6D-4F92-49C4-9C4E-A24200BF9DE6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{77F256F5-BB49-4D7C-8D6C-5976D162727A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{77F256F5-BB49-4D7C-8D6C-5976D162727A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{77F256F5-BB49-4D7C-8D6C-5976D162727A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{77F256F5-BB49-4D7C-8D6C-5976D162727A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D333E4E3-4070-4BC3-89C9-EA4423FDF8B9}
+	EndGlobalSection
+EndGlobal
