--- src/IdentificationNumber/Entities/NumberType.cs
@@ -1,9 +1,0 @@
-﻿namespace IdentificationNumber.Models
-{
-    public enum NumberType
-    {
-        Unknown = 0,
-        Person = 1,
-        Coordination = 2
-    }
-}
