--- src/IdentificationNumber/Enums/Gender.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace IdentificationNumber.Enums
-{
-    public enum Gender
-    {
-        Female = 0,
-        Male = 1
-    }
-}
