--- src/IdentificationNumber/Enums/PersonType.cs
@@ -1,9 +1,0 @@
-﻿namespace IdentificationNumber.Enums
-{
-    public enum PersonType
-    {
-        Unknown,
-        Person,
-        Coordination
-    }
-}
