+++ src/IdentificationNumber/Interfaces/IIdentificationNumber.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace IdentificationNumber.Interfaces
+﻿namespace psafth.IdentificationNumber.Interfaces
 {
     public interface IIdentificationNumber
     {
