+++ src/psafth.IdentificationNumber.Swedish/Entities/Gender.cs
@@ -1,0 +1,8 @@
+﻿namespace psafth.IdentificationNumber.Swedish.Entities
+{
+    public enum Gender
+    {
+        Female = 0,
+        Male = 1
+    }
+}
