+++ src/psafth.IdentificationNumber.Swedish/Entities/PersonNumberType.cs
@@ -1,0 +1,9 @@
+﻿namespace psafth.IdentificationNumber.Entities
+{
+    public enum PersonNumberType
+    {
+        Unknown = 0,
+        Person = 1,
+        Coordination = 2
+    }
+}
