+++ src/psafth.IdentificationNumber.Swedish/Helpers/CommonRegex.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace IdentificationNumber.Helpers
+namespace psafth.IdentificationNumber.Swedish.Helpers
 {
     public static class CommonRegex
     {
