+++ src/psafth.IdentificationNumber.Swedish/psafth.IdentificationNumber.Swedish.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\IdentificationNumber\psafth.IdentificationNumber.csproj" />
+  </ItemGroup>
+
+</Project>
