--- src/swedish-identification-number.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33103.184
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IdentificationNumber", "IdentificationNumber\IdentificationNumber.csproj", "{46E35B14-2DBB-4603-885B-A84A9D71BFC9}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IdentificationNumber.Tests", "IdentificationNumber.Tests\IdentificationNumber.Tests.csproj", "{A06DF01A-8C46-410F-A10D-A97544E06FE5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IdentificationNumber.Examples", "IdentificationNumber.Examples\IdentificationNumber.Examples.csproj", "{CB3087F6-70E0-41D9-A87F-025382234A8C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{46E35B14-2DBB-4603-885B-A84A9D71BFC9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A06DF01A-8C46-410F-A10D-A97544E06FE5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CB3087F6-70E0-41D9-A87F-025382234A8C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D333E4E3-4070-4BC3-89C9-EA4423FDF8B9}
-	EndGlobalSection
-EndGlobal
